split the tests into 3 files since the current test file was getting a little long. Added tests for redo.
